+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 * Focused Rage (thanks wang429)
 * Reap and Sow (thanks LankSSBM)
 * Old Book (thanks LankSSBM)
+* Possession (thanks nic3245)
 * Repressed Memory (thanks nic3245)
 
 ### Wanderer Updates